Reece

+++ Assets/Scripts/Stats/Modifier/AdditiveModifier.cs
@@ -4,10 +4,10 @@
 public class AdditiveModifier : Modifier
 {
     public override int Order { get; } = 10000;
-    public AdditiveModifier(double value) : base(value) { }
-    public AdditiveModifier(double value, StatType statType) : base(value, statType) { }
+    public AdditiveModifier(float value) : base(value) { }
+    public AdditiveModifier(float value, StatType statType) : base(value, statType) { }
 
-    public override double Calculate(double value)
+    public override float Calculate(float value)
     {
         return value * (this.value as dynamic);
     }
